fix: code comments corrected
